# api/openapi/平台接口规范.yaml
openapi: 3.1.0
info:
  title: 基础能力平台 API
  version: 0.1.0-draft
  description: |
    基础能力平台的前后端接口草案。实现基线为 Go 模块化单体、MySQL、JWT Cookie 会话，
    所有时间均为 UTC RFC 3339 格式。本文件与 docs/前后端接口契约.md 共同作为联调契约；
    表结构和领域约束以 docs/核心模型与MySQL设计.md、docs/系统架构.md 为准。
servers:
  - url: /api/v1
    description: 平台 API 相对地址
...
security:
  - cookieAuth: []
...
paths:
...
  /applications:
    get:
      tags: [Application Registry]
      summary: 分页查询接入应用
      operationId: listApplications
      x-permission: "platform:application:read"
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Keyword' }
        - { name: status, in: query, schema: { type: string, enum: [ACTIVE, DISABLED] } }
      responses:
        '200': { $ref: '#/components/responses/ApplicationList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      tags: [Application Registry]
      summary: 创建接入应用
      operationId: createApplication
      x-permission: "platform:application:create"
      x-audit-action: platform:application:create
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ApplicationCreateRequest' } }
      responses:
        '201': { $ref: '#/components/responses/Application' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /applications/{application_id}:
    parameters:
      - { $ref: '#/components/parameters/ApplicationId' }
    get:
      tags: [Application Registry]
      summary: 查询接入应用详情
      operationId: getApplication
      x-permission: "platform:application:read"
      responses:
        '200': { $ref: '#/components/responses/Application' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Application Registry]
      summary: 更新接入应用
      description: 应用 code 创建后不可修改；必须提交当前 version 以进行乐观锁控制。
      operationId: updateApplication
      x-permission: "platform:application:update"
      x-audit-action: platform:application:update
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ApplicationUpdateRequest' } }
      responses:
        '200': { $ref: '#/components/responses/Application' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/VersionConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: bp_session
      description: JWT 会话 Cookie。实际名称由 IAM_SESSION_COOKIE_NAME 配置，默认 bp_session。
    applicationAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 由 /oauth2/token 使用 Client Credentials 签发的接入应用 JWT。接收端校验 Ed25519（EdDSA）签名、issuer、audience、token_use=application 与时间声明，并以数据库当前启用的 OAuth Client、Application、Environment 和 scope 复核令牌绑定；审计上报必须具备 audit.ingest 范围。
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 本 OIDC 发行方在 Authorization Code 或 Refresh Token 授权中签发的终端用户 Access Token。
    clientCredentialsBasic:
      type: http
      scheme: basic
      description: OAuth 令牌端点的 client_secret_basic 客户端认证。
  parameters:
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Keyword: { name: keyword, in: query, schema: { type: string, maxLength: 100 } }
...
    ApplicationId: { name: application_id, in: path, required: true, schema: { $ref: '#/components/schemas/ULID' } }
...
  responses:
    Unauthenticated:
      description: 未登录、会话已过期或应用凭据无效。
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Forbidden:
      description: 当前主体缺少权限或数据范围不允许。
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: 资源不存在，或当前主体无权感知该资源。
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Conflict:
      description: 唯一约束、状态转换或业务冲突。
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    VersionConflict:
      description: 乐观锁冲突；前端应刷新数据后重新提交。
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    ValidationError:
      description: 请求字段校验失败。
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    Application: { description: 接入应用对象, content: { application/json: { schema: { $ref: '#/components/schemas/ApplicationEnvelope' } } } }
    ApplicationList: { description: 接入应用分页结果, content: { application/json: { schema: { $ref: '#/components/schemas/ApplicationListEnvelope' } } } }
...
  schemas:
    ULID: { type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$', examples: ['01J123456789ABCDEFGHJKMNPQ'] }
    Timestamp: { type: string, format: date-time, description: UTC RFC 3339 时间戳。 }
    Version: { type: integer, minimum: 1, description: 聚合根乐观锁版本号。 }
...
    Envelope:
      type: object
      required: [code, message, request_id, data]
      properties:
        code: { type: string, const: OK }
        message: { type: string, examples: ['操作成功'] }
        request_id: { $ref: '#/components/schemas/ULID' }
        data: {}
...
    ErrorEnvelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          examples: [AUTH_UNAUTHENTICATED, AUTH_FORBIDDEN, IAM_VERSION_CONFLICT, PLATFORM_CAPABILITY_NOT_ENABLED]
        message: { type: string }
        request_id: { $ref: '#/components/schemas/ULID' }
        details: { type: object, additionalProperties: true }
    PageData:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: {} }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
    Application:
      type: object
      required: [application_id, code, name, application_type, status, version, created_at, updated_at]
      properties:
        application_id: { $ref: '#/components/schemas/ULID' }
        code: { type: string, maxLength: 64, pattern: '^[a-z][a-z0-9._-]*$' }
        name: { type: string, maxLength: 128 }
        application_type: { type: string, enum: [spa, web, backend, mobile, third_party] }
        owner_org_id: { $ref: '#/components/schemas/ULID', nullable: true }
        owner_user_id: { $ref: '#/components/schemas/ULID', nullable: true }
        homepage_url: { type: string, format: uri, nullable: true }
        description: { type: string, maxLength: 1000, nullable: true }
        status: { type: string, enum: [DRAFT, ACTIVE, SUSPENDED, RETIRED] }
        version: { $ref: '#/components/schemas/Version' }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
    ApplicationCreateRequest:
      type: object
      additionalProperties: false
      required: [code, name, application_type, status]
      properties:
        code: { type: string, maxLength: 64, pattern: '^[a-z][a-z0-9._-]*$' }
        name: { type: string, minLength: 1, maxLength: 128 }
        application_type: { type: string, enum: [spa, web, backend, mobile, third_party] }
        owner_org_id: { $ref: '#/components/schemas/ULID', nullable: true }
        owner_user_id: { $ref: '#/components/schemas/ULID', nullable: true }
        homepage_url: { type: string, format: uri, nullable: true }
        description: { type: string, maxLength: 1000, nullable: true }
        status: { type: string, enum: [DRAFT, ACTIVE, SUSPENDED, RETIRED] }
    ApplicationUpdateRequest:
      type: object
      additionalProperties: false
      required: [name, application_type, status, version]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        application_type: { type: string, enum: [spa, web, backend, mobile, third_party] }
        owner_org_id: { $ref: '#/components/schemas/ULID', nullable: true }
        owner_user_id: { $ref: '#/components/schemas/ULID', nullable: true }
        homepage_url: { type: string, format: uri, nullable: true }
        description: { type: string, maxLength: 1000, nullable: true }
        status: { type: string, enum: [DRAFT, ACTIVE, SUSPENDED, RETIRED] }
        version: { $ref: '#/components/schemas/Version' }
...
    ApplicationEnvelope: { allOf: [ { $ref: '#/components/schemas/Envelope' }, { type: object, properties: { data: { $ref: '#/components/schemas/Application' } } } ] }
    ApplicationListEnvelope: { allOf: [ { $ref: '#/components/schemas/Envelope' }, { type: object, properties: { data: { $ref: '#/components/schemas/ApplicationPage' } } } ] }
...
    ApplicationPage: { allOf: [ { $ref: '#/components/schemas/PageData' }, { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Application' } } } } ] }
...